scripts/requirements-yolo.txt
ultralytics>=8.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pillow>=10.0.0
